+++ zerver/lib/event_schema.py
@@ -1067,6 +1067,7 @@ realm_user_type = DictType(
         ("timezone", str),
         ("date_joined", str),
         ("email_address_visibility", EnumType(UserProfile.EMAIL_ADDRESS_VISIBILITY_TYPES)),
+        ("delivery_email", OptionalType(str)),
     ]
 )
 
@@ -1140,7 +1141,7 @@ realm_user_person_types = dict(
         required_keys=[
             # vertical formatting
             ("user_id", int),
-            ("delivery_email", str),
+            ("delivery_email", OptionalType(str)),
         ],
     ),
     email=DictType(